+++ assets/js/admin/course-edit.js
@@ -1,35 +1,24 @@
 /**
  * WordPress dependencies
  */
-import { select, subscribe, dispatch, useSelect } from '@wordpress/data';
-import { applyFilters } from '@wordpress/hooks';
+import { select, subscribe } from '@wordpress/data';
 import domReady from '@wordpress/dom-ready';
 import { registerPlugin } from '@wordpress/plugins';
-import { __ } from '@wordpress/i18n';
-import {
-	PluginDocumentSettingPanel,
-	PluginSidebar,
-	PluginSidebarMoreMenuItem,
-} from '@wordpress/edit-post';
-import { Slot } from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
 import { startBlocksTogglingControl } from './blocks-toggling-control';
-import CourseTheme from './course-theme';
-import CourseVideoSidebar from './course-video-sidebar';
-import CoursePricingPromoSidebar from './course-pricing-promo-sidebar';
-import CourseAccessPeriodPromoSidebar from './course-access-period-promo-sidebar';
 import {
 	extractStructure,
 	getFirstBlockByName,
 } from '../../blocks/course-outline/data';
-
-import SenseiIcon from '../../icons/logo-tree.svg';
-
-const pluginSidebarHandle = 'sensei-lms-course-settings-sidebar';
-const pluginDocumentHandle = 'sensei-lms-document-settings-sidebar';
+import {
+	CourseSidebar,
+	SenseiSettingsDocumentSidebar,
+	pluginSidebarHandle,
+	pluginDocumentHandle,
+} from './course-settings-plugin-sidebar';
 
 ( () => {
 	const editPostSelector = select( 'core/edit-post' );
@@ -119,63 +108,9 @@ domReady( () => {
  * Plugins
  */
 
-const CourseSidebar = () => {
-	const hideCoursePricing = applyFilters( 'senseiCoursePricingHide', false );
-	const hideAccessPeriod = applyFilters(
-		'senseiCourseAccessPeriodHide',
-		false
-	);
-	return (
-		<>
-			<PluginSidebarMoreMenuItem
-				target={ pluginSidebarHandle }
-				icon={ <SenseiIcon height="20" width="20" color="#43AF99" /> }
-			>
-				{ __( 'Sensei Settings', 'sensei-lms' ) }
-			</PluginSidebarMoreMenuItem>
-			<PluginSidebar
-				name={ pluginSidebarHandle }
-				title={ __( 'Sensei Settings', 'sensei-lms' ) }
-				icon={ <SenseiIcon color="#43AF99" /> }
-				position={ 1 }
-			>
-				{ ! hideCoursePricing && <CoursePricingPromoSidebar /> }
-				{ ! hideAccessPeriod && <CourseAccessPeriodPromoSidebar /> }
-				<Slot name="SenseiCourseSidebar" />
-				<CourseTheme />
-				<CourseVideoSidebar />
-			</PluginSidebar>
-		</>
-	);
-};
-
 registerPlugin( pluginSidebarHandle, {
 	render: CourseSidebar,
 } );
-
-const SenseiSettingsDocumentSidebar = () => {
-	const isSenseiEditorPanelOpen = useSelect( ( select ) => {
-		return select( 'core/edit-post' ).isEditorPanelOpened(
-			`${ pluginDocumentHandle }/${ pluginDocumentHandle }`
-		);
-	} );
-	if ( isSenseiEditorPanelOpen ) {
-		// when 'Sensei Settings' is clicked, isSenseiEditorPanelOpen returns true, so we open the 'Sensei Settings'
-		// plugin sidebar and then close the 'Sensei Settings' panel which sets isSenseiEditorPanelOpen back to false.
-		dispatch( 'core/edit-post' ).openGeneralSidebar(
-			`${ pluginSidebarHandle }/${ pluginSidebarHandle }`
-		);
-		dispatch( 'core/edit-post' ).toggleEditorPanelOpened(
-			`${ pluginDocumentHandle }/${ pluginDocumentHandle }`
-		);
-	}
-	return (
-		<PluginDocumentSettingPanel
-			name={ pluginDocumentHandle }
-			title={ __( 'Sensei Settings ', 'sensei-lms' ) }
-		></PluginDocumentSettingPanel>
-	);
-};
 
 registerPlugin( pluginDocumentHandle, {
 	render: SenseiSettingsDocumentSidebar,
